# Timbangan/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Timbangan Digital IoT API",
        "description": "API untuk integrasi perangkat timbangan IoT (Arduino/ESP8266) dengan sistem penimbangan digital. Semua endpoint menggunakan token perangkat untuk autentikasi.",
        "contact": {
            "name": "Timbangan Digital Team"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": " Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/incoming-singkong/settings": {
            "get": {
                "tags": [
                    "Incoming Singkong"
                ],
                "summary": "Get pengaturan sesi aktif timbangan Incoming Singkong",
                "description": "Mendapatkan informasi sesi aktif untuk penimbangan singkong masuk, termasuk info supplier dan no surat.",
                "operationId": "61a7422582822a92c0a20e8002000e2d",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "DEV-TOKEN-SINGKONG-001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "ready",
                                                "idle"
                                            ],
                                            "example": "ready"
                                        },
                                        "kode_produksi": {
                                            "type": "string",
                                            "example": "KP-SINGKONG-001"
                                        },
                                        "nama_produk": {
                                            "type": "string",
                                            "example": "Singkong Basah"
                                        },
                                        "operator": {
                                            "type": "string",
                                            "example": "Budi"
                                        },
                                        "no_surat": {
                                            "type": "string",
                                            "example": "SJ-001/2024"
                                        },
                                        "nama_supplier": {
                                            "type": "string",
                                            "example": "PT. Singkong Jaya"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/incoming-singkong/weight": {
            "post": {
                "tags": [
                    "Incoming Singkong"
                ],
                "summary": "Kirim data berat timbangan Incoming Singkong",
                "description": "Menyimpan data berat singkong masuk. Membutuhkan sesi aktif yang sudah diinput oleh operator.",
                "operationId": "f9373a10246307e065e4e279db272737",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "token",
                                    "weight"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "DEV-TOKEN-SINGKONG-001"
                                    },
                                    "weight": {
                                        "description": "Berat dalam kg",
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berat berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WeightSuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Berat tidak valid / Tidak ada sesi aktif",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "WeightSuccessResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Data berat 25.5 kg tersimpan"
                    },
                    "record_id": {
                        "type": "integer",
                        "example": 42
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            }
        }
    },
...
}